moved AI files to respective areas
Moved AI files to respective areas
Removed duplicate physics file
Slightly changed player speed & removed basic gui menu
