+++ resources/scripts/_MasterCreation.lua
@@ -19,7 +19,7 @@ end
 addGameObject("scripts/CatPhysics.lua");
 addGameObject("scripts/CatPhysics2.lua");
 addGameObject("scripts/Water.lua")
-addGameObject("scripts/guiBasic.lua")
+--addGameObject("scripts/guiBasic.lua")
 addGameObject("scripts/pauseMenuSetup.lua")
 addGameObject("scripts/helpMenuSetup.lua")
 addGameObject("scripts/credits_setup.lua")
