+++ src/Controller/AI/Counter.cpp
