+++ src/Controller/GameAssetFactory.hpp
@@ -18,10 +18,10 @@
 #include "Model/GameObject/PointLightObject.hpp"
 #include "Model/GameObject/SpotLightObject.hpp"
 #include "Controller/Terrain/TexturedTerrain.hpp"
-#include "AI/singletons.h"
+#include "Model/singletons.h"
 #include "Model/GameObject/Projectile.hpp"
 #include "Model/GameObject/NetworkedItem.hpp"
-#include "AI/NPC.hpp"
+#include "Model/GameObject/NPC.hpp"
 /**
  * @class GameAssetFactory
  * @brief A parameterized factory singleton for creating GameObjects.
