+++ src/Model/singletons.h
@@ -1,21 +1,18 @@
 #pragma once
 
 #include "singleton.h"
-#include "entityManager.h"
-#include "dynamicEnum.h"
-#include "messageDispatcher.h"
-#include "world.hpp"
-#include "Counter.hpp"
+#include "Controller/AI/entityManager.h"
+#include "Controller/AI/messageDispatcher.h"
+#include "Controller/Pathfinding/world.hpp"
+#include "Controller/AI/Counter.hpp"
 #include "Controller/Physics/physicsEntityManager.h"
 
-typedef singleton<dynamicEnum> message;
 typedef singleton<entityManager> entityMan;
 typedef singleton<messageDispatcher> messageMan;
 typedef singleton<world> theWorld;
 typedef singleton<Counter> idManager;
 typedef singleton<physicsEntityManager> physicsEntityMan;
 
-#define fsmMessage message::Instance()
 #define entityMgr entityMan::Instance()
 #define messageMgr messageMan::Instance()
 #define gameWorld theWorld::Instance()
